can take in photo input but not output yet
Got a test page with a form to take in a form with an image then have it converted into buffer form that MongoDB accepts and creates a document with.
Only did it for creation, not update.
Static files to test it are in the photoTtest folder with testForm in the name.